refactor(remote-control): 拆分文件访问边界
