+++ pinvou3-app/src-tauri/src/features/remote_control/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod file_access;
 pub(crate) mod manager;
 mod platform;
 mod protocol;
@@ -5,6 +6,8 @@ mod relay_client;
 
 pub use manager::{RelaySettingsInfo, RemoteControlManager};
 pub use protocol::{WebAccessInfo, WebAccessStatus};
+
+pub(crate) const MAX_TRANSFER_CHUNK_BYTES: usize = 256 * 1024;
 
 use rand::distr::Alphanumeric;
 use rand::Rng;
